Remove duplicate code for package info
